unit tests(WIP)

+++ url-shorter-backend/src/lib.rs
@@ -3,8 +3,8 @@ pub mod utils;
 pub mod web;
 
 use lazy_static::lazy_static;
-use std::sync::atomic::AtomicU64;
+use std::sync::atomic::{AtomicU32, AtomicU64};
 
 lazy_static! {
-    pub static ref SHORT_URL_COUNT: AtomicU64 = AtomicU64::new(0);
+    pub static ref SHORT_URL_COUNT: AtomicU32 = AtomicU32::new(0);
 }
